add system for tracer bullets and fix frame dependent movement

+++ src/components/Animation.hpp
@@ -5,7 +5,7 @@
 #include "utils/Vec.hpp"
 
 struct Animation {
-    uint64_t startTime; // Current animation time since it started
+    float startTime; // Current animation time since it started
     uint64_t duration; // In ms
     uint16_t curStep; // Current step of the animation
     uint16_t steps; // Number of steps of the animation
